refactor code pertaining to comments to new component
